% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LH_Priors.R
\name{Add_Correlation_Taylor}
\alias{Add_Correlation_Taylor}
\title{Add correlation structure to selected values}
\usage{
Add_Correlation_Taylor(
  simulated.data.trunc,
  cor_matrix = NULL,
  Var_Names = NULL,
  truncSD = 1.96
)
}
\arguments{
\item{simulated.data.trunc}{A data.frame returned by \code{Sample_Trunc_LogNormal_Taylor}}

\item{cor_matrix}{A correlation matrix returned by \code{calc_Correlation_Matrix}}

\item{Var_Names}{A data.frame with the required variable names from
\code{calc_Correlation_Matrix} in the first column, and matching variable names from
\code{Sample_Trunc_LogNormal_Taylor} in the second column}

\item{truncSD}{The number of standard deviations to truncate}
}
\value{
A data.frame with values from \code{simulated.data.trunc} updated
with the correlation structure from \code{cor_matrix}
}
\description{
Add correlation structure to selected values
}
