% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LH_Priors.R
\name{Calculate_Steepness}
\alias{Calculate_Steepness}
\alias{iVB}
\title{Calculate Steepness using the Mangel (2010) method}
\usage{
Calculate_Steepness(x, DF, plus.group = 16)

iVB(t0, K, Linf, L)
}
\arguments{
\item{x}{sample number (up to \code{nrow(DF)})}

\item{DF}{A data.frame with \code{nrow} samples and specifically named columns, see \code{Details}}

\item{plus.group}{Age of the plus group}
}
\value{
the predicted steepness value
}
\description{
Calculate Steepness using the Mangel (2010) method
}
\details{
\code{DF} must have the following column names:
\itemize{
\item \code{mort}: Natural mortality rate
\item \code{vLinf}:
\item \code{vonK}:
\item \code{vtto}:
\item \code{length.weight.alpha}:
\item \code{length.weight.beta}:
\item \code{L50}:
\item \code{L50_95}:
\item \code{BFConstant}:
\item \code{Bfa}:
\item \code{BFb}:
\item \code{ParChange}:
\item \code{RealizedFecFactor}:
\item \code{NS}:
\item \code{Se}:
}
}
\section{Functions}{
\itemize{
\item \code{iVB()}: Calculate inverse von Bert

}}
