% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LH_Priors.R
\name{get_FishLife_Predictions}
\alias{get_FishLife_Predictions}
\alias{calc_Correlation_Matrix}
\title{Extract predicted values from FishLife}
\usage{
get_FishLife_Predictions(Genus = "Xiphias", Species = "gladius", level = 2)

calc_Correlation_Matrix(Cov_pred)
}
\arguments{
\item{Genus}{Genus}

\item{Species}{Species}

\item{level}{Prediction level to extract: 1:}

\item{Cov_pred}{\code{Cov_pred} object returned by \code{get_FishLife_Predictions}}
}
\value{
A list object returned from \code{FishLife::Plot_taxa}
}
\description{
Extract predicted values from FishLife
}
\section{Functions}{
\itemize{
\item \code{calc_Correlation_Matrix()}: Calculate the correlation matrix

}}
