% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LH_Priors.R
\name{Sample_Trunc_LogNormal_Taylor}
\alias{Sample_Trunc_LogNormal_Taylor}
\title{Sample Life History Parameters for the Taylor Priors}
\usage{
Sample_Trunc_LogNormal_Taylor(
  n = 10000,
  truncSD = 1.96,
  path = "Resources/Taylor_Values.csv"
)
}
\arguments{
\item{n}{The number of samples}

\item{truncSD}{The number of standard deviations to truncate}

\item{path}{The full file path for the CSV file with the \code{Mean} and \code{SD} values. Assumes
the first column are the names of the parameters}
}
\value{
A data.frame of the sampled values
}
\description{
Generates \code{n} samples from a truncated log-normal distribution from \code{Mean} and
\code{CV} values located in the CSV specified in \code{path}
}
